Removes jade
Uses weback.DefinePlugin to provide config values
