--- index-builder.js
@@ -1,23 +1,0 @@
-var jade = require( 'jade' );
-
-function getConfig() {
-	var config = {};
-
-	try {
-		config = require( './config.js' )();
-	} catch ( e ) {
-		console.warn( 'no configuration file present' );
-	}
-
-	return config;
-};
-
-function getTemplateContent( templateParams ) {
-	return jade.renderFile( './views/app.jade', {
-		pkg: require( './package.json' ),
-		o: templateParams,
-		config: getConfig()
-	} );
-}
-
-module.exports = getTemplateContent;
