+++ package.json
@@ -24,6 +24,7 @@
     "npm": "3.10.10"
   },
   "devDependencies": {
+    "add-asset-html-webpack-plugin": "^1.0.2",
     "appcache-webpack-plugin": "1.3.0",
     "autoprefixer": "6.5.4",
     "babel-core": "6.21.0",
@@ -41,8 +42,6 @@
     "eslint-loader": "1.6.1",
     "eslint-plugin-react": "6.8.0",
     "html-webpack-plugin": "2.24.1",
-    "jade": "1.11.0",
-    "jade-loader": "0.8.0",
     "json-loader": "0.5.4",
     "node-sass": "4.1.1",
     "postcss-loader": "1.2.1",
