--- views/app.jade
@@ -1,15 +1,0 @@
-doctype html
-html(manifest=o.htmlWebpackPlugin.files.manifest)
-  head
-    meta(charset="utf-8")
-    title= o.htmlWebpackPlugin.options.title || 'Simplenote'
-    //- iOS webviews do not support indexedDB!
-    //- meta(name="apple-mobile-web-app-capable" content="yes")
-    meta(name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no")
-    link(rel="icon" href="/favicon.ico")
-    link(rel="stylesheet" type="text/css" href="https://fonts.googleapis.com/css?family=Source+Sans+Pro:300,300italic,400,400italic,600,600italic")
-    script(type="text/javascript")
-      != 'var config=' + JSON.stringify( Object.assign( { version: pkg.version }, config ) ) + ';'
-  body
-    script(src="vendor.dll.js")
-    script(src="app.js")
